+++ front_end/src/i18n/locales/en.ts
@@ -251,8 +251,10 @@ export default {
             modeRecord: ' mode record: '
         },
         videos: 'All videos',
-        exportArbiterCSV: 'Export as stats_csv.csv',
-        exportArbiterCSVTooltip: 'Compatible with the datasheet generated by Minesweeper Arbiter.<br/> Not supporting Leff, Reff, Deff, GZiNi and HZiNi at present.',
+        exportJSON: 'Export as JSON',
+        exportJSONTooltip: 'Raw data fetched from the server.',
+        exportArbiterCSV: 'Export as CSV',
+        exportArbiterCSVTooltip: 'Compatible with <span style="font-family: \'Courier New\', Courier, monospace;">stats_csv.csv</span> generated by Minesweeper Arbiter.<br/> Not supporting Leff, Reff, Deff, GZiNi and HZiNi at present.',
         upload: {
             title: 'Video Upload',
             dragOrClick: `Drag files here or <em>click here to select</em>`,
